Tracing: create empty metadata if none is present.
When getting the metadata from the context fails, there is no metadata
passed. This means that when we try to create headers from the metadata,
where we set metadata information, it will try and write to a nil map,
which causes a panic.

This initialises an empty metadata object when it can't create one from
context.